CT/COVNet/requirements.txt
batchgenerators==0.19.7
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imageio==2.8.0
joblib==0.14.1
kiwisolver==1.1.0
kornia==0.2.0
linecache2==1.0.0
matplotlib==3.2.1
networkx==2.4
nibabel==3.0.2
numpy==1.18.2
pandas==1.0.3
Pillow==7.0.0
protobuf==3.11.3
pydicom==1.4.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
SimpleITK==1.2.4
six==1.14.0
tensorboardX==2.0
threadpoolctl==2.0.0
torch==1.4.0
torchsummary==1.5.1
torchvision==0.5.0
traceback2==1.4.0
unittest2==1.1.0
xlrd==1.2.0
